Ghosts are more prone to go from one side to its opposite
Previous version of the A*-algorithm was not very good in dealing with paths going from one side of the board to its opposite side.
The evaluation of 'possible distance' between two opposite locations was far too high compared to other measurements. For these reasons ghosts were not prone to use these 'shortcuts'.